+++ first_ray/assignment_1.hpp
@@ -10,7 +10,7 @@ struct ass_1_renderer
     // TODO
     // Convert from recursive to iterative
     Vector3f Li(const ray &r, Scene *scene, const int &depth, const hit_record &prev_hrec,
-        const float &prev_bsdf_pdf, sampler &random_sampler);
+        const double &prev_bsdf_pdf, sampler &random_sampler);
 
     void Render(Scene *scene, viewer *film_viewer, tf::Taskflow &tf);
