+++ first_ray/box.h
@@ -9,8 +9,8 @@ class box : public hitable
 public:
     box() {}
     box(const Vector3f &pmin, const Vector3f &pmax, material *mat);
-    virtual bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const;
-    virtual bool bounding_box(float t0, float t1, aabb &box) const
+    virtual bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const;
+    virtual bool bounding_box(double t0, double t1, aabb &box) const
     {
         box = aabb(pmin, pmax);
         return true;
@@ -33,7 +33,7 @@ box::box(const Vector3f &pmin, const Vector3f &pmax, material *mat) : pmin(pmin)
     list_ptr = new hitable_list(list, 6);
 }
 
-bool box::hit(const ray &r, float t_min, float t_max, hit_record &rec) const
+bool box::hit(const ray &r, double t_min, double t_max, hit_record &rec) const
 {
     return list_ptr->hit(r, t_min, t_max, rec);
 }
