+++ first_ray/hitable_list.cpp
@@ -1,7 +1,7 @@
 #include "hitable_list.h"
 #include "util.h"
 
-bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) const
+bool hitable_list::hit(const ray& r, double t_min, double t_max, hit_record& rec) const
 {
 	hit_record temp_rec;
 	bool hit_anything = false;
@@ -9,7 +9,7 @@ bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) 
 
 	for (int i = 0; i < list_size; i++)
 	{
-		if (list[i]->hit(r, t_min, (float)closest_so_far, temp_rec))
+		if (list[i]->hit(r, t_min, (double)closest_so_far, temp_rec))
 		{
 			hit_anything = true;
 			closest_so_far = temp_rec.t;
@@ -20,7 +20,7 @@ bool hitable_list::hit(const ray& r, float t_min, float t_max, hit_record& rec) 
 	return hit_anything;
 };
 
-bool hitable_list::bounding_box(float t0, float t1, aabb& box) const
+bool hitable_list::bounding_box(double t0, double t1, aabb& box) const
 {
 	if (list_size < 1) return false;
 	aabb temp_box;
@@ -43,10 +43,10 @@ bool hitable_list::bounding_box(float t0, float t1, aabb& box) const
 	return true;
 }
 
-float hitable_list::pdf_direct_sampling(const hit_record& lrec, const Vector3f& to_light) const
+double hitable_list::pdf_direct_sampling(const hit_record& lrec, const Vector3f& to_light) const
 {
-	float weight = 1.0 / list_size;
-	float sum = 0;
+	double weight = 1.0 / list_size;
+	double sum = 0;
 	for (int i = 0; i < list_size; i++)
 		sum += weight * list[i]->pdf_direct_sampling(lrec, to_light);
 
@@ -55,13 +55,13 @@ float hitable_list::pdf_direct_sampling(const hit_record& lrec, const Vector3f& 
 
 Vector3f hitable_list::sample_direct(hit_record& rec, const Vector3f& o, const Vector2f &sample) const
 {
-	float rand = sample[0];
+	double rand = sample[0];
 	int index = int(rand * list_size);
 	if (index == list_size) index -= 1;
 	return list[index]->sample_direct(rec, o, sample);
 }
 
-int hitable_list::pick_sample(const float& sample) const
+int hitable_list::pick_sample(const double& sample) const
 {
 	int index = int(sample * list_size);
 	if (index == list_size) index -= 1;
